Implement stats/release
This will grab the latest release of RN for our marketing page.
