+++ package.json
@@ -39,7 +39,7 @@
     "ejs": "^2.3.1",
     "express": "^4.12.3",
     "express-session": "^1.10.4",
-    "github": "^0.2.4",
+    "github": "git+https://github.com/reviewninja/node-github.git",
     "glob": "^5.0.3",
     "less-middleware": "^2.0.1",
     "merge": "^1.2.0",
